+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   the `Session.End()` method and only for the active session.
 - `IdentifyUser` can be called with `null` or an empty string.  
   This simulates a log off event and no longer re-applies the user tag on split sessions.
+- Improve handling of client IP in combination with server-side detection.
 
 ## 2.0.0 [Release date: 2020-06-24]
 [GitHub Releases](https://github.com/Dynatrace/openkit-dotnet/releases/tag/v2.0.0)